+++ src/Utils/ServerInfo.java
@@ -2,14 +2,16 @@ package Utils;
 
 import java.io.Serializable;
 import java.net.Socket;
+import java.sql.Array;
+import java.util.ArrayList;
 
 public class ServerInfo implements Serializable {
     private int MAXIMUM, CONECTIONSAMOUNT;
-    public Socket[] clientsSockets;
+    public ArrayList<Socket> clientsSockets;
 
     public ServerInfo(){}
 
-    public ServerInfo(int MAXIMUM, int CONECTIONSAMOUNT, Socket[] clientsSockets) {
+    public ServerInfo(int MAXIMUM, int CONECTIONSAMOUNT, ArrayList<Socket> clientsSockets) {
         this.MAXIMUM = MAXIMUM;
         this.CONECTIONSAMOUNT = CONECTIONSAMOUNT;
         this.clientsSockets = clientsSockets;
@@ -31,15 +33,15 @@ public class ServerInfo implements Serializable {
         this.CONECTIONSAMOUNT = CONECTIONSAMOUNT;
     }
 
-    public Socket[] getClientsSockets() {
+    public ArrayList<Socket> getClientsSockets() {
         return clientsSockets;
     }
 
-    public void setClientsSockets(Socket[] clientsSockets) {
+    public void setClientsSockets(ArrayList<Socket> clientsSockets) {
         this.clientsSockets = clientsSockets;
     }
 
     public void addClientSocket(Socket clientSocket, int position) {
-        this.clientsSockets[position] = clientSocket;
+        this.clientsSockets.add(clientSocket);
     }
 }
